+++ glyph/src/glyph.rs
@@ -1,8 +1,4 @@
-use crate::PxScale;
-
-/// An (x, y) coordinate. `Point { x: f32, y: f32 }`
-pub type Point = ab_glyph_rasterizer::Point;
-pub use ab_glyph_rasterizer::point;
+use crate::{Point, PxScale};
 
 /// Glyph id.
 ///
@@ -63,4 +59,113 @@ pub struct Glyph {
     pub scale: PxScale,
     /// Position of this glyph.
     pub position: Point,
+}
+
+/// Old version of [`v2::GlyphImage`].
+#[deprecated(since = "0.2.22", note = "Deprecated in favor of `v2::GlyphImage`")]
+#[derive(Debug, Clone)]
+pub struct GlyphImage<'a> {
+    /// Offset of the image from the normal origin (top at the baseline plus
+    /// ascent), measured in pixels at the image's current scale.
+    pub origin: Point,
+    /// Current scale of the image in pixels per em.
+    pub scale: f32,
+    /// Raw image data, not a bitmap in the case of [`GlyphImageFormat::Png`] format.
+    pub data: &'a [u8],
+    /// Format of the raw data.
+    pub format: GlyphImageFormat,
+}
+
+pub mod v2 {
+    use crate::{GlyphImageFormat, Point};
+
+    /// A pre-rendered image of a glyph, usually used for emojis or other glyphs
+    /// that can't be represented only using an outline.
+    #[non_exhaustive]
+    #[derive(Debug, Clone)]
+    pub struct GlyphImage<'a> {
+        /// Offset of the image from the normal origin (top at the baseline plus
+        /// ascent), measured in pixels at the image's current scale.
+        pub origin: Point,
+        /// Image width.
+        ///
+        /// It doesn't guarantee that this value is the same as set in the `data` in the case of
+        /// [`GlyphImageFormat::Png`] format.
+        pub width: u16,
+        /// Image height.
+        ///
+        /// It doesn't guarantee that this value is the same as set in the `data` in the case of
+        /// [`GlyphImageFormat::Png`] format.
+        pub height: u16,
+        /// Pixels per em of the selected strike.
+        pub pixels_per_em: u16,
+        /// Raw image data, see [`format`](GlyphImageFormat).
+        pub data: &'a [u8],
+        /// Format of the raw [`data`](Self::data).
+        pub format: GlyphImageFormat,
+    }
+}
+
+/// Valid formats for a [`GlyphImage`].
+// Possible future formats: SVG, JPEG, TIFF
+#[non_exhaustive]
+#[derive(Debug, Clone)]
+pub enum GlyphImageFormat {
+    Png,
+
+    /// A monochrome bitmap.
+    ///
+    /// The most significant bit of the first byte corresponds to the top-left pixel, proceeding
+    /// through succeeding bits moving left to right. The data for each row is padded to a byte
+    /// boundary, so the next row begins with the most significant bit of a new byte. 1 corresponds
+    /// to black, and 0 to white.
+    BitmapMono,
+
+    /// A packed monochrome bitmap.
+    ///
+    /// The most significant bit of the first byte corresponds to the top-left pixel, proceeding
+    /// through succeeding bits moving left to right. Data is tightly packed with no padding. 1
+    /// corresponds to black, and 0 to white.
+    BitmapMonoPacked,
+
+    /// A grayscale bitmap with 2 bits per pixel.
+    ///
+    /// The most significant bits of the first byte corresponds to the top-left pixel, proceeding
+    /// through succeeding bits moving left to right. The data for each row is padded to a byte
+    /// boundary, so the next row begins with the most significant bit of a new byte.
+    BitmapGray2,
+
+    /// A packed grayscale bitmap with 2 bits per pixel.
+    ///
+    /// The most significant bits of the first byte corresponds to the top-left pixel, proceeding
+    /// through succeeding bits moving left to right. Data is tightly packed with no padding.
+    BitmapGray2Packed,
+
+    /// A grayscale bitmap with 4 bits per pixel.
+    ///
+    /// The most significant bits of the first byte corresponds to the top-left pixel, proceeding
+    /// through succeeding bits moving left to right. The data for each row is padded to a byte
+    /// boundary, so the next row begins with the most significant bit of a new byte.
+    BitmapGray4,
+
+    /// A packed grayscale bitmap with 4 bits per pixel.
+    ///
+    /// The most significant bits of the first byte corresponds to the top-left pixel, proceeding
+    /// through succeeding bits moving left to right. Data is tightly packed with no padding.
+    BitmapGray4Packed,
+
+    /// A grayscale bitmap with 8 bits per pixel.
+    ///
+    /// The first byte corresponds to the top-left pixel, proceeding through succeeding bytes
+    /// moving left to right.
+    BitmapGray8,
+
+    /// A color bitmap with 32 bits per pixel.
+    ///
+    /// The first group of four bytes corresponds to the top-left pixel, proceeding through
+    /// succeeding pixels moving left to right. Each byte corresponds to a color channel and the
+    /// channels within a pixel are in blue, green, red, alpha order. Color values are
+    /// pre-multiplied by the alpha. For example, the color "full-green with half translucency"
+    /// is encoded as `\x00\x80\x00\x80`, and not `\x00\xFF\x00\x80`.
+    BitmapPremulBgra32,
 }